Split the `dep` commands into separate files
