+++ ddev/changelog.d/17256.fixed
@@ -1,0 +1,1 @@
+Split the `dep` commands into separate files
